Get PnMPI version from git.
Instead of managing the PnMPI version by hand in the root CMakeLists.txt file,
the version will be get from git now. If get is not available, it will be read
from a file named '.version' in the project root directory. This file should
be created just before packing a tarball, but not be versioned.