Splitted PnMPI constructors.
PnMPI uses separate functions for different constructors now. If gcc's construc-
tors are not available, a fallback constructor will be used. In addition, all
constructors will use the provided argc and argv parameters instead of using
the read_cmdline function. If the fallback constructor must be used, it will
take care on getting argc and argv.