PnMPI threading level may be changed via ENV.
The default MPI threading level is the highest available level. For performance-
sensitive applications, this may slow down them, especially if they are not
multithreaded but are using the app hooks. To decrease the MPI overhead, the
threading level for MPI may be adjusted by the 'PNMPI_THREADING_LEVEL' variable.