New header organisation.
In preparation for the debug object library two new functions have been added
for CMake: pnmpi_add_header and pnmpi_add_private_header. These will copy the
defined header files to a temporary location as if they are installed at the
system. This should make the developement processes more simple, as the
developers don't have to take care anymore if the header is installed or not.

A common situation for this is, if header a/a.h wants to include ../b/b.h, but
after installation they are in the same directory 'headers'. Now a simply
includes 'pnmpi/b.h' - equal if its installed or used at compile time.