New debug message functions.
The new PnMPI debug functions provide an easy to use way to print debug
messages, warnings or errors without the need to initialize these with other
macros - everything will be initialized on demand at the first time of using
them. Special care was taken on C89 compability in pedantic mode.