Added new macro to install man pages.
Doxygen generated man pages are not really smart, but with Doxygen modules you
may get some structure in the wild. A new macro helps to install the man page
and its symlinks.

All you have to do is defining a new Doxygen module by \defgroup and attach the
symbols to it with \ingroup. In CMake you need to install the man pages with
pnmpi_add_man(<GROUP NAME> <SYMBOL NAME>). All man pages will be installed in
the man3 section.