Replaced the code for printing the PnMPI header.
Instead of printing the PnMPI header and some information about the used config
in the MPI_Init* functions, the new pnmpi_print_banner function will be used. It
will be called on app_startup hooks now, too.

Instead of printing detailed information about the used config in the header,
these will be hidden now, as the user may enable the debug level on demand now.