Replaced PMPI_Barrier by MPI_Barrier in print_banner.
Some PnMPI modules (like GTI) manipulate communicators to have ranks not
participating in the application but on other tasks. In this case a barrier on
MPI_COMM_WORLD may fail, if the not-participating ranks hang in MPI_Init for
a long time.
To solve this issue, MPI_Barrier will be used, so it gets wrapped by the module,
which will change the communicator to the internal one for the participating
ranks.