Fixed testcases and better error reporting.
There has been a deadlocks in test-mpi-c and test-mpi-fortran, as all ranks sent
at the same time but none was receiving. Now all ranks send their message to
rank 0 which is deadlock free.

A new header file provides macros for MPI_* calls, so the return value has not
to be checked in any testcase, but the wrapper macros will check it.