Renamed pnmpi_status to PNMPI_status_t.
As other typenames of the PnMPI service interface are named PNMPI_*_t, this one
was named in this scheme, too.

In the glibc coding standard typenames ending with _t are not allowed, but as
the PNMPI interface was already named in this scheme, a renaming of the other
typenames was not possible for compatibility reasons. As all typenames have the
PNMPI_ prefix, a collision should be impossible.