Added optional versions of app hooks.
New versions of the app hooks give ensure, that a module can execute code as
late as possible depending on the environment. This gives the ability to run the
same module in different environments whre constructors might not be available.
The hook will then be executed in MPI_Init / MPI_Finalize.