Fixed interface detection for static PnMPI.
If PnMPI is linked statically against an executable, the executable will have a
symbol named MPI_Init, so the interface detection will always asume the applica-
tion uses the C MPI interface, even if the application is Fortran or doesn't use
any MPI functions.

The new behaviour is to disable the interface detection via the nm command, if
PnMPI is compiled statically. This results in an error, if PnMPI can't detect
the used interface for pnmpi_app_startup, so the user has to define the inter-
face via environment in this case.