All app-hooks are optional now.
Instead of handling two different hooks for optional and non-optional versions
of the app hooks, only the optional version will be used now.

If there is a need to force the execution of the hook before main, this may be
done by checking pnmpi_constructor_called of src/pnmpi/constructor.c by a
service function. As there is no need for this feature at the moment, this is
not implemented yet.