Reworked hook invocation.
Instead of calling the hooks in all modules, a recursion like those for the
other MPI wrapper functions was implemented. A module has to call
PNMPI_Service_CallHook at the point of recursion, now. This gives the ability
to control, where the other module's hooks will be called in the execution.